requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
gunicorn>=20.1.0          
psycopg2-binary>=2.9       
django-cors-headers>=4.0   
djangorestframework-simplejwt>=5.3  
python-decouple>=3.8       
drf-yasg>=1.21.7   
python-dotenv
       